// nestjs-quiz-app/src/modules/quiz/dto/create.quiz.dto.ts
import { Category } from 'src/modules/category/category.entity';
import { Question } from 'src/modules/question/question.entity';

export class CreateQuizDto {
  title: string;

  description: string;

  maxMarks: number;

  numberOfQuestion: number;

  active: boolean;

  category: Category;

  question: Question[];
}